Moved UpdateInfo into cluster/manual/_utils.py.
