--- phy/cluster/manual/_update_info.py
@@ -1,31 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""UpdateInfo class."""
-
-#------------------------------------------------------------------------------
-# Imports
-#------------------------------------------------------------------------------
-
-from ...utils._misc import Bunch
-
-
-#------------------------------------------------------------------------------
-# UpdateInfo class
-#------------------------------------------------------------------------------
-
-def update_info(**kwargs):
-    """Hold information about clustering changes."""
-    d = dict(
-        description=None,  # information about the update: 'merge', 'assign',
-                           # or 'metadata_<name>'
-        spikes=[],  # all spikes affected by the update
-        added=[],  # new clusters
-        deleted=[],  # deleted clusters
-        descendants=[],  # pairs of (old_cluster, new_cluster)
-        metadata_changed=[]  # clusters with changed metadata
-    )
-    d.update(kwargs)
-    return Bunch(d)
-
-
-UpdateInfo = update_info
